# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Gestión de Pistas",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/"
        }
    ],
    "paths": {
...
        "/api/get-bookings": {
            "get": {
                "tags": [
                    "Reservas"
                ],
                "summary": "Obtener todas las reservas registradas",
                "operationId": "1b9c2ebc3b43604f8812ab74c7049b74",
                "responses": {
                    "202": {
                        "description": "\n    *           $bookings (Object [])"
                    },
                    "403": {
                        "description": "\n    *           Necesitas ser Administrador para realizar esta operación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/add-reserve": {
            "post": {
                "tags": [
                    "Reservas"
                ],
                "summary": "Registra una nueva reserva",
                "operationId": "d97f1d47cdad0beefa95962b5eb8a45a",
                "parameters": [
                    {
                        "name": "horaInicio",
                        "in": "query",
                        "description": "Define a la hora que empieza la reserva",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "horaFinalizacion",
                        "in": "query",
                        "description": "Define a la hora que acaba la reserva",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "fechaCita",
                        "in": "query",
                        "description": "Espcifica la fecha de la reserva",
                        "required": true,
                        "schema": {
                            "type": "date"
                        }
                    },
                    {
                        "name": "numLista",
                        "in": "query",
                        "description": "Define el número en la lista en las pitas que tengan un aforo y no un horario",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "users_id",
                        "in": "query",
                        "description": "ID del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "pistas_id",
                        "in": "query",
                        "description": "ID de la pista",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "\n    *           Reserva añadida con éxito"
                    },
                    "400": {
                        "description": "\n    *           Falló validación\n    *           Esta reserva debe tener un número de lista y no un horario\n    *           Esta reserva debe tener un horario y no un número de lista\n    *           La pista sobre la que se ha hecho la reserva no existe"
                    },
                    "404": {
                        "description": "\n    *           Esta pista no está disponible"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/delete-reserve/{id}": {
            "delete": {
                "tags": [
                    "Reservas"
                ],
                "summary": "Eliminar una reserva",
                "operationId": "9f86a1e9f95bfee4cfd9de75781bad49",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la reserva",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "\n    *           Se ha eliminado la reserva correctamente"
                    },
                    "400": {
                        "description": "\n    *           Esta reserva no existe"
                    },
                    "403": {
                        "description": "\n    *           Necesitas ser Administrador o dueño de esta reserva para realizar esta operación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/exists-reserve/{court_id}/{user_id}": {
            "get": {
                "tags": [
                    "Reservas"
                ],
                "summary": "Comprueba si existe una reserva",
                "operationId": "764d5ea93731c7f46bd79d103e44776a",
                "parameters": [
                    {
                        "name": "court_id",
                        "in": "path",
                        "description": "ID de la pista",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "ID del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "\n    *           True"
                    },
                    "404": {
                        "description": "\n    *           False"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/get-booking-user/{user_id}": {
            "get": {
                "tags": [
                    "Reservas"
                ],
                "summary": "Devuelve todas las reservas de un usuario",
                "operationId": "e8e68cd32c809901aa4f0a256def55d2",
                "parameters": [
                    {
                        "name": "users_id",
                        "in": "path",
                        "description": "ID del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "\n    *           $bookings (Object [])\n    *           No tienes reservas"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/get-booking-court/{court_id}": {
            "get": {
                "tags": [
                    "Reservas"
                ],
                "summary": "Devuelve todas las reservas que tiene una pista",
                "operationId": "c9ab68ad4aacb3587233ada6bf26d57a",
                "parameters": [
                    {
                        "name": "court_id",
                        "in": "path",
                        "description": "ID de la pista",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "\n    *           $bookings (Object [])\n    *           No tienes reservas"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "apiKey",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "http",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost/oauth/authorize",
                        "tokenUrl": "http://localhost/oauth/token",
                        "refreshUrl": "http://localhost/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    }
}